Día3/requirements.txt
camelcase==0.2
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-MySQLdb==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
python-dotenv==0.21.0
Werkzeug==2.2.2
